Title: 소수의 거듭제곱, Time: 232ms, Memory: 15MB, Status: Wrong Answer - CodeTree

+++ 240729/소수의 거듭제곱/power-of-prime-number.java
@@ -7,8 +7,8 @@ public class Main {
         StringTokenizer st;
 
         st = new StringTokenizer(br.readLine());
-        long a = Integer.parseInt(st.nextToken());
-        long b = Integer.parseInt(st.nextToken());
+        long a = Long.parseLong(st.nextToken());
+        long b = Long.parseLong(st.nextToken());
         int max = (int)Math.pow(b, 0.5);
 
         boolean[] isPrime = new boolean[max+1];
